weekly_news/requirements.txt
feedparser>=6.0.0
google-generativeai>=0.8.0
requests>=2.31.0
